Add comments (and fix a typo on new line 12)
The original regular expression included "\s*", which matches any amount of whitespace, including no whitespace.  I changed it to "\s+", which matches at least one, but possibly more, whitespace characters.

With the data given, this still works because every line includes whitespace, but if there were a line that included only digits with no space, then $b would be the final digit, and $a would be all other digits.